refactor: split form inputs and programmatic validation
Appears that zod doesn't provide a nice way to pipe
different inputs into the same verifier. `.pipe()` short-circuits
on errors in the first object and doesn't get error on unrelated
fields if they are in the second verifier. So extract constraints
into a plain object and then pipe each field separately.